Moved parser service to graduation_imitator
