+++ mse1h2025_imitate/graduation_imitator/dto/PresentationData.py
