--- mse1h2025_imitate/parser/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class ParserConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'parser'
