+++ mse1h2025_imitate/users_manager/admin.py
@@ -1,3 +1,7 @@
 from django.contrib import admin
 
-# Register your models here.
+# Register you
+# models here.
+
+
+
